# docs/swagger.yaml
definitions:
  internal_handler.ActiveCallRequest:
    properties:
      actions:
        description: action 列表
        items:
          $ref: '#/definitions/shortpress-server_internal_plugins.ActionParams'
        type: array
      hook:
        description: hook 类型
        type: string
    required:
    - actions
    - hook
    type: object
  internal_handler.InstallPluginRequest:
    properties:
      config:
        description: 插件配置（可选，JSON格式）
        type: string
      plugin_id:
        description: 插件ID
        type: string
    required:
    - plugin_id
    type: object
  internal_handler.RegisterPluginRequest:
    properties:
      author:
        description: 插件作者
        type: string
      description:
        description: 插件描述
        type: string
      name:
        description: 插件名称
        type: string
      type:
        description: 插件类型：payment, content, analytics 等
        type: string
      version:
        description: 插件版本
        type: string
    required:
    - name
    - type
    type: object
  internal_handler.UninstallPluginRequest:
    properties:
      plugin_id:
        description: 插件ID
        type: string
    required:
    - plugin_id
    type: object
...
  shortpress-server_internal_api.Response:
    properties:
      code:
        type: integer
      data: {}
      info:
        type: string
    type: object
...
  shortpress-server_internal_plugins.ActionParams:
    properties:
      action:
        type: string
      params:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: shortpress
    url: http://www.swagger.io/support
  description: This is shortpress server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Shortpress Server API
  version: "0.8"
paths:
...
  /api/plugins/hook/active_call:
    post:
      consumes:
      - application/json
      description: 插件通过此接口主动调用指定的 action
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      - description: Plugins ID
        in: header
        name: X-Plugins-Id
        required: true
        type: string
      - description: 请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ActiveCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回 action 执行结果
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      summary: 插件主动调用接口
      tags:
      - plugins
  /api/plugins/install:
    post:
      consumes:
      - application/json
      description: 为当前站点安装指定插件
      parameters:
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      - description: 安装插件请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.InstallPluginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 安装成功，返回密钥
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "404":
          description: 插件不存在
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      summary: 安装插件到站点
      tags:
      - plugin-management
  /api/plugins/installed/list:
    get:
      consumes:
      - application/json
      description: 获取当前站点已安装的插件列表
      parameters:
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      summary: 获取站点已安装插件列表
      tags:
      - plugin-management
  /api/plugins/list:
    get:
      consumes:
      - application/json
      description: 获取所有已注册的插件列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      summary: 获取可用插件列表
      tags:
      - plugin-management
  /api/plugins/register:
    post:
      consumes:
      - application/json
      description: 在系统中注册新插件
      parameters:
      - description: 注册插件请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RegisterPluginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      summary: 注册插件
      tags:
      - plugin-management
  /api/plugins/uninstall:
    post:
      consumes:
      - application/json
      description: 从当前站点卸载指定插件
      parameters:
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      - description: 卸载插件请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UninstallPluginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 卸载成功
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "404":
          description: 插件未安装
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      summary: 卸载站点插件
      tags:
      - plugin-management
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"